rename : eventTag => eventStatusTag
