+++ src/components/common/EventCard/EventCard.tsx
@@ -4,7 +4,7 @@ import { getTimeStamp } from '@/utils/getTimeStamp';
 import { useNavigate } from 'react-router-dom';
 
 import Avatar from '../Avatar/Avatar';
-import EventStatusTag from '../EventTag/EventTag';
+import EventStatusTag from '../EventStatusTag/EventTag';
 import Poster from '../Poster/Poster';
 import {
   ClubInfoWrapperStyle,
